moved delete station in to helper folder
