+++ lib/functions/dts-process.js
@@ -5,16 +5,7 @@ const invokeLambda = require('../helpers/invoke-lambda')
 const Joi = require('joi')
 const axios = require('axios')
 const { HTTP_NOT_FOUND } = require('../constants')
-
-async function deleteStation (stationId) {
-  try {
-    await pg('station_display_time_series').where({ station_id: stationId }).delete()
-    logger.info(`Deleted data for RLOI id ${stationId}`)
-  } catch (error) {
-    logger.error(`Error deleting data for station ${stationId}`, error)
-    throw error
-  }
-}
+const deleteStation = require('../helpers/imtd-api').deleteStation
 
 async function insertStation (stationDataArray) {
   try {
